Refactor test classes into packages
