+++ backend/src/test/java/edu/zut/bookrider/integration/service/NavigationServiceIntegrationTest.java
@@ -1,47 +1,24 @@
-package edu.zut.bookrider.unit;
+package edu.zut.bookrider.integration.service;
 
 import edu.zut.bookrider.dto.CoordinateDTO;
 import edu.zut.bookrider.dto.NavigationResponseDTO;
 import edu.zut.bookrider.exception.InvalidCoordinatesException;
 import edu.zut.bookrider.service.NavigationService;
 import edu.zut.bookrider.service.enums.TransportProfile;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.web.client.RestTemplate;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.when;
 
 @SpringBootTest
-public class NavigationServiceTest {
+public class NavigationServiceIntegrationTest {
 
     @Autowired
     private NavigationService navigationService;
-
-    @Mock
-    private RestTemplate restTemplate;
-
-    private String validRouteJson;
-    private String noRouteJson;
-    private String invalidCoordinatesJson;
-
-    @BeforeEach
-    void setUp() throws IOException {
-        validRouteJson = new String(Files.readAllBytes(Paths.get("src/test/resources/navigationServiceTest/valid-route-api-response.json")));
-        noRouteJson = new String(Files.readAllBytes(Paths.get("src/test/resources/navigationServiceTest/no-route-api-response.json")));
-        invalidCoordinatesJson = new String(Files.readAllBytes(Paths.get("src/test/resources/navigationServiceTest/invalid-coordinates-api-response.json")));
-    }
 
     @Test
     void whenPossibleRouteFromInput_thenNavigationStepsShouldBeCorrect() {
@@ -55,8 +32,6 @@ public class NavigationServiceTest {
         CoordinateDTO endCoordinates = new CoordinateDTO(endLatitude, endLongitude);
 
         TransportProfile transportProfile = TransportProfile.CAR;
-
-        when(restTemplate.getForObject(anyString(), eq(String.class))).thenReturn(validRouteJson);
 
         NavigationResponseDTO navigationResponseDTO = navigationService.getDirectionsFromCoordinates(startCoordinates, endCoordinates, transportProfile);
 
@@ -99,8 +74,6 @@ public class NavigationServiceTest {
         CoordinateDTO endCoordinates = new CoordinateDTO(14.504721, 53.434444);
         TransportProfile transportProfile = TransportProfile.CAR;
 
-        when(restTemplate.getForObject(anyString(), eq(String.class))).thenReturn(noRouteJson);
-
         InvalidCoordinatesException thrownException = assertThrows(
                 InvalidCoordinatesException.class,
                 () -> navigationService.getDirectionsFromCoordinates(startCoordinates, endCoordinates, transportProfile)
@@ -114,8 +87,6 @@ public class NavigationServiceTest {
         CoordinateDTO startCoordinates = new CoordinateDTO(14.504721, 53.434444);
         CoordinateDTO endCoordinates = new CoordinateDTO(290.504721, 53.434444);
         TransportProfile transportProfile = TransportProfile.CAR;
-
-        when(restTemplate.getForObject(anyString(), eq(String.class))).thenReturn(invalidCoordinatesJson);
 
         InvalidCoordinatesException thrownException = assertThrows(
                 InvalidCoordinatesException.class,
